backend/requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
django-localflavor
django-crispy-forms
django-jquery
django-bootstrap5
fontawesomefree
djangorestframework
django-environ